backend/requirements/base.txt
attrs==18.2.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
Django==2.2.12
django-environ==0.4.5
django-redis==4.9.0
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
mccabe==0.6.1
mypy==0.620
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pygraphviz==1.5.0
boto3==1.9.64

Pygments==2.2.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
rope==0.11.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1

django-extensions==2.2.9
djangorestframework==3.10.3
djangorestframework-gis==0.14
djangorestframework-mvt==0.2.2
django-geojson==2.12.0
argon2-cffi==18.3.0
django-allauth==0.38.0
fiona==1.8.4
shapely==1.6.4.post2
django-grappelli==2.12.1
django-storages==1.7.1
django-model-utils==3.1.2
django-sendfile==0.3.11
django-rest-auth==0.9.3
django-filter==2.2.0
GDAL==2.1.3
geopandas==0.7.0
django-pandas==0.6.1
django-watson==1.5.3
